Change "i" to "intent", elongate introscreens with 1 second, move protocol package into the network package.
