+++ EscapeApp/app/src/main/java/com/context/kroket/escapeapp/minigames/B_TapGame.java
@@ -145,12 +145,11 @@ public class B_TapGame extends AppCompatActivity {
 
     super.onStart();
 
-    Intent i = new Intent(this, ConnectionService.class);
-    bindService(i, mConnection, Context.BIND_AUTO_CREATE);
+    Intent intent = new Intent(this, ConnectionService.class);
+    bindService(intent, mConnection, Context.BIND_AUTO_CREATE);
 
     // Change the current activity to this
     ((ActivityManager) this.getApplicationContext()).setCurrentActivity(this);
-
   }
 
   /**
@@ -207,8 +206,8 @@ public class B_TapGame extends AppCompatActivity {
       }
     } else {
       clickButton.setText("Finish");
-      Intent i = new Intent(this, WaitingActivity.class);
-      startActivity(i);
+      Intent intent = new Intent(this, WaitingActivity.class);
+      startActivity(intent);
     }
   }
