+++ EscapeApp/app/src/main/java/com/context/kroket/escapeapp/network/protocol/CommandParser.java
@@ -1,4 +1,4 @@
-package com.context.kroket.escapeapp.protocol;
+package com.context.kroket.escapeapp.network.protocol;
 
 import java.util.ArrayList;
 import java.util.HashMap;
